improved fork(), exec*(), posix_spawn*() handling

+++ Nyxian/LindChain/ProcEnvironment/Server/ServerProtocol.h
@@ -33,7 +33,7 @@
 /*
  posix_spawn
  */
-- (void)spawnProcessWithPath:(NSString*)path withArguments:(NSArray*)arguments withEnvironmentVariables:(NSDictionary *)environment withMapObject:(FDMapObject*)mapObject withReply:(void (^)(unsigned int))reply;
+- (void)spawnProcessWithPath:(NSString*)path withArguments:(NSArray*)arguments withEnvironmentVariables:(NSDictionary *)environment withMapObject:(FDMapObject*)mapObject withReply:(void (^)(int64_t))reply;
 
 /*
  Process Info
